==> 08_isoform.Rmd <==
---
title: "Isoform"
author: "Yuting Xie"
date: "2022/11/8"
output: html_document
---
```{r}
rm(list = ls())
pacman::p_load(readxl,magrittr,
               RColorBrewer,
               Rtsne, umap,
               pheatmap,vioplot,
               ggpubr, ggplot2,
               corrplot, stringr,reshape2,preprocessCore,data.table)
```
#data
```{r}
setwd("D:/1_guomics_xyt/1Project/hba_research_workflow/9_DataAnalysis_20220901/Reactome/tissue grouped reactome pathway")
pro_info <- data.table::fread("FIleInfo_HBA_20220816_withoutFALSE_and_pool.csv",data.table = F)
# pro_info<-data.table::fread("E:/HBA/cor_pro_RNA/FIleInfo_HBA_20220816_withoutFALSE_and_pool.csv",data.table = F)
setwd("D:/1_guomics_xyt/1Project/hba_research_workflow/9_DataAnalysis_20220901/WGCNA_analysis")
pro<-data.table::fread('HBA_prot_matrix_normolized_20220811.csv', stringsAsFactors = F, data.table = F, fill = T)
rownames(pro)<-pro[,1]
pro<-pro[,-1]
pro<-as.data.frame(t(pro))

#Q9UPP5 has been merged into Q5XKL5
rownames(pro)[grep("Q9UPP5",rownames(pro))]<-"Q5XKL5-2"
pro1<-pro[grep("-",rownames(pro)),]
pro1$group<-sapply(rownames(pro1),function(x){str_split(x,"-")[[1]][1]})
IsoPro<-unique(pro1$group)
pro1<-pro[unlist(sapply(IsoPro, function(x){grep(x, rownames(pro))})), ]

pro1<-as.data.frame(t(pro1))

pro1<-pro1[match(pro_info$sample[which(pro_info$pathology=="health",)],rownames(pro1)),]
pro1$W_G<-pro_info$W_G[match(rownames(pro1),pro_info$sample)]
pro2<-pro1[grep("G",pro1$W_G),] 

pro3<-pro2[,-ncol(pro2)]
pro3<-pro3[,order(colnames(pro3))]
pro4<-pro3         #[,apply(pro3,2,function(x){sum(!is.na(x))>0})]
```
#isoform number
```{r}
proIso<-data.frame(Iso=colnames(pro4))
proIso$Iso<-sapply(proIso$Iso,function(x){str_split(x,"-")[[1]][1]})
proIso2<-as.data.frame(table(proIso$Iso))
colnames(proIso2)<-c("pro","isoNumHBA")

fasta<-data.table::fread("D:/1_guomics_xyt/1Project/hba_research_workflow/9_DataAnalysis_20220901/isoform/uniprot-download_true_fields_accession_2Creviewed_2Cid_2Cprotein_nam-2022.11.08-07.57.42.24.tsv",data.table = F)
fasta<-fasta[,c(1,8)]
fasta1<-fasta[order(fasta$`Alternative products (isoforms)`,decreasing = T),]
fasta1[,2]<-sapply(fasta1[,2],function(x){str_split(x,"isoforms=")[[1]][2]})
fasta2<-na.omit(fasta1)
fasta2[,2]<-sapply(fasta2[,2],function(x){str_split(x,";")[[1]][1]})
colnames(fasta2)<-c("pro","isoNumFasta")

datRatio<-merge(proIso2,fasta2,by="pro")
datRatio[,2]<-as.numeric(datRatio[,2])
datRatio[,3]<-as.numeric(datRatio[,3])
datRatio$ratio<-round(datRatio$isoNumHBA/datRatio$isoNumFasta,2)
datRatio<-datRatio[order(datRatio$ratio,decreasing = T),]
datRatio<-data.frame(datRatio,row.names = "pro")

# setwd("D:/1_guomics_xyt/1Project/hba_research_workflow/9_DataAnalysis_20220901/isoform")
# write.csv(datRatio,"isoform_ratio.csv")
```
#figure for an isoform
```{r}
datFig<-pro4[,grep("Q15700",colnames(pro4))]
datFig<-datFig[apply(datFig,1,function(x){sum(!is.na(x))>0}),]
datFig$BBA<-pro_info$BBA_Anatomical.and.modified.Cyto.architectonic.descriptions[match(rownames(datFig),pro_info$sample)]
datFig1<-melt(datFig,id="BBA")
datFig2<-as.data.frame(reshape2::dcast(datFig1, BBA~variable, mean, na.rm=T))
rownames(datFig2)<-datFig2[,1]
datFig2<-datFig2[,-1]

datFig2<-log(datFig2,base=50)
datFig2[is.na(datFig2)]<-0

datFig3<-as.data.frame(t(datFig2))
# datFig3<-as.data.frame(apply(datFig3, 1, function(x){scale(x)}))
# rownames(datFig3)<-rownames(datFig2)
# datFig3[datFig3<0]<-0
# datFig3<-as.data.frame(t(datFig3))

datFig3$isoform<-rownames(datFig3)
datFig4<-melt(datFig3)

ggplot(datFig4, aes(x = variable, y = isoform, height = value, group = isoform)) +
  geom_ridgeline(fill="lightblue") +
  theme(axis.text.x=element_text(angle=90,hjust=1,size=3))
setwd("D:/1_guomics_xyt/1Project/hba_research_workflow/9_DataAnalysis_20220901/isoform")
ggsave("Q15700_isoform_geom_rideline_log50_value.pdf",width=12,height=8)

```
#heatmap fot all isoform
```{r}
datHeatmap<-pro4[,order(colnames(pro4))]
datHeatmap$BBA<-pro_info$BBA_Anatomical.and.modified.Cyto.architectonic.descriptions[match(rownames(datHeatmap),pro_info$sample)]
datHeatmap1<-reshape2::melt(datHeatmap, by="BBA")
datHeatmap2<-reshape2::dcast(datHeatmap1, BBA~variable, mean, na.rm=T)
rownames(datHeatmap2)<-datHeatmap2[,1]
datHeatmap2<-datHeatmap2[,-1]

datHeatmap3<-datHeatmap2  #[,apply(datHeatmap2,2,function(x){sum(!is.na(x))>0})]  #0.5*nrow(datHeatmap2)

# datHeatmap3[is.na(datHeatmap3)]<-0

annotation_col<-data.frame(pro=colnames(datHeatmap3),   isoform=sapply(colnames(datHeatmap3),function(x){str_split(x,"-")[[1]][1]}))
isoformNum<-as.data.frame(table(annotation_col$isoform))

info <- pro_info[,c(18:21)]
info <- na.omit(unique(info))
info <- info[order(info$Order),]

SP<-isoformNum[isoformNum$Freq>1,]  #377
i <- which(SP=="Q15149")
for(i in 1:nrow(SP)){
  datHeatmap4<-datHeatmap3[,unlist(sapply(SP$Var1[i], function(x){grep(x, colnames(datHeatmap3))}))]

#---------mean by Lobe--------------------
  max_Q15149 <- max(datHeatmap4[,1])
  datHeatmap4[datHeatmap4>max_Q15149] <- max_Q15149
  datHeatmap4$Lobe <- info$Lobe[match(rownames(datHeatmap4), info$BBA_Anatomical.and.modified.Cyto.architectonic.descriptions)]
  datMean <- reshape2::melt(datHeatmap4, id="Lobe")
  datMean1 <- acast(datMean, Lobe~variable, mean)
  datMean1 <- t(datMean1)
  datMean1 <- datMean1[c(1,4,3,2), ]
  setwd("D:/1_guomics_xyt/1Project/hba_research_workflow/9_DataAnalysis_20220901/isoform/select pathway heatmap")
  pheatmap(datMean1,
         fontsize=8,
         fontsize_row=10,
         fontsize_col = 10,
         border_color="white",
         color = colorRampPalette(c("white", "#eeeacd", "#258881"))(100),   #"#6A7F67", "firebrick4"
         # annotation_col = annotation_col2,
         # annotation_row = annotation_row,
         #annotation_colors = list(W_G=c(rep("#4472C4",26),rep("#EE781C",19),Region=rainbow(45))),
         # na_color = "gray",
         trace = "none", 
         scale = "none", 
         cluster_rows = F,
         cluster_cols = F,
         show_rownames = T, 
         show_colnames = T, 
         filename = paste(SP$Freq[i], SP$Var1[i],seq="_","isoform_heatmap_mean_by_Lobe.pdf"),       
         main = paste(SP$Freq[i], SP$Var1[i],seq="_","isoform mean expression"),
         height = 5, width =10
  )
#-----------------------------------------  
  datHeatmap5<-datHeatmap4[info$BBA_Anatomical.and.modified.Cyto.architectonic.descriptions, ]
  rownames(datHeatmap5) <- as.factor(rownames(datHeatmap5))
  library(forcats)
  rownames(datHeatmap5) <- fct_inorder(rownames(datHeatmap5))
  
  datHeatmap5 <- as.data.frame(t(datHeatmap5))
  max_Q15149 <- max(datHeatmap5[1, ])
  datHeatmap5[datHeatmap5>max_Q15149] <- max_Q15149
  datHeatmap5 <- datHeatmap5[c(1,4,3,2), ]
  # datHeatmap5 <- 2^(datHeatmap5)
  # datHeatmap5 <- log10(datHeatmap5)
  #as.data.frame(apply(datHeatmap4,2,function(x){scale(x)}))
  # rownames(datHeatmap5)<-rownames(datHeatmap4)
  # datHeatmap5[is.na(datHeatmap5)]<-0
  setwd("D:/1_guomics_xyt/1Project/hba_research_workflow/9_DataAnalysis_20220901/isoform/select pathway heatmap")
  pheatmap(datHeatmap5,
         fontsize=8,
         fontsize_row=10,
         fontsize_col = 10,
         border_color="white",
         color = colorRampPalette(c("white", "#d2e7d7", "#6796b5"))(100),   #"#6A7F67", "firebrick4"
         # annotation_col = annotation_col2,
         # annotation_row = annotation_row,
         #annotation_colors = list(W_G=c(rep("#4472C4",26),rep("#EE781C",19),Region=rainbow(45))),
         # na_color = "gray",
         trace = "none", 
         scale = "none", 
         cluster_rows = F,
         cluster_cols = F,
         show_rownames = T, 
         show_colnames = T, 
         filename = paste(SP$Freq[i], SP$Var1[i],seq="_","isoform_heatmap_log10_max3.pdf"),       
         main = paste(SP$Freq[i], SP$Var1[i],seq="_","isoform expression"),
         height = 5, width =15
  )
}

# annotation_row<-data.frame(BBA=rownames(datHeatmap3), Lobe=pro_info$Lobe[match(rownames(datHeatmap3),pro_info$BBA_Anatomical.and.modified.Cyto.architectonic.descriptions)])
# annotation_row<-annotation_row[order(rownames(annotation_row)),]
# annotation_row<-data.frame(annotation_row,row.names = "BBA")
# 
# datHeatmap3<-datHeatmap3[rownames(annotation_row),]
# 
# annotation_col2<-data.frame(isoform=sapply(colnames(datHeatmap3),function(x){str_split(x,"-")[[1]][1]}))
# 
# setwd("D:/1_guomics_xyt/1Project/hba_research_workflow/9_DataAnalysis_20220901/isoform")
# pheatmap(datHeatmap3,
#          fontsize_row=15,
#          fontsize_col = 8,
#          color = colorRampPalette(c("white","#6A7F67","firebrick4"))(100), 
#          annotation_col = annotation_col2,
#          annotation_row = annotation_row,
#          #annotation_colors = list(W_G=c(rep("#4472C4",26),rep("#EE781C",19),Region=rainbow(45))),
#          # na_color = "gray",
#          trace = "none", 
#          scale = "none", 
#          cluster_rows = F,
#          cluster_cols = F,
#          show_rownames = T, 
#          show_colnames = T, 
#          filename = "isoform_heatmap.pdf",       
#          main = "All isoform expression",
#          height = 30,width =30)

```
