==> 17_Differentially expressed proteins between brain regions_Grey_matter.Rmd <==
---
title: "Differentially expressed proteins between brain regions_pheatmap"
author: "Yuting Xie"
date: "2023/01/11"
output: html_document
---
rm(list = ls())
pacman::p_load(readxl,magrittr,
               RColorBrewer,
               Rtsne, umap,
               pheatmap,vioplot,
               ggpubr, ggplot2,
               corrplot, stringr,reshape2,preprocessCore)

#data
```{r}
setwd("D:/1_guomics_xyt/1Project/hba_research_workflow/9_DataAnalysis_20220901/differential analysis of brain regions_20230111")
info<-read.csv("FIleInfo_HBA_20220816_withoutFALSE_and_pool.csv",header = T,sep = ",",stringsAsFactors = F)
# info$SampleName<-info$sample_ID
# pacman::p_load(stringr)
# info$SampleName<-sapply(info$SampleName,function(x){str_split(x,"_")[[1]][2]})
# info$SampleName<-paste(info$batch_ID,info$SampleName,sep="_")
# table(info$Lobe)

df<-data.table::fread("HBA_prot_matrix_normolized_20220811.csv", data.table = F)
rownames(df)<-df[,1]
df<-df[,-1]

df1<-df[-grep("pool",rownames(df)), ] #delet pools
df1$pathology <- info$pathology[match(rownames(df1), info$sample)]
df2<-df1[grep("health", df1$pathology), ] #delet epilepsy
df3 <- df2[, -ncol(df2)]
df3<-as.data.frame(t(df3))

df3$pro<-sapply(rownames(df3),function(x){str_split(x,"-")[[1]][1]})
df4<-melt(df3,id="pro")
df5<-reshape2::dcast(df4, pro~variable, mean, na.rm=T)
#d3$pro[duplicated(d3$pro)]
df5<-data.frame(df5,row.names = "pro")
df5 <- as.data.frame(t(df5))

df5$W_G<-info$W_G[match(rownames(df5),info$sample)]
dfG<-df5[grep("G", df5$W_G),]
dfW <- df5[grep("W", df5$W_G),]

setwd("D:/1_guomics_xyt/1Project/hba_research_workflow/9_DataAnalysis_20220901/differential analysis of brain regions_20230111")
save(dfG, file = "matrix_G.RData")
save(dfW, file = "matrix_W.RData")
# df4$Lobe<-info$Lobe[match(df4$rep,info$SampleName)]
# #delet Lobe$"\","Cerebellum","SubcorticalNuclei","NA"   ("InsularLobe" don't delet)
# 
# df5<-df4[-which(is.na(df4$Lobe)),]
# df6<-df5[-grep("Cerebellum",df5$Lobe),]
# df7<-df6[-grep("SubcorticalNuclei",df6$Lobe),]
# # df8<-df7[-grep("InsularLobe",df7$Lobe),]
# # df8<-df7[-grep("\\\\",df7$Lobe),]
# df8<-df7
# 
# df8$BBA<-info$BBA_Anatomical.and.modified.Cyto.architectonic.descriptions[match(df8$rep,info$SampleName)]
# df8$region<-paste(df8$BBA,df8$Lobe,sep="--")
```

#Fold change
```{r}
setwd("W:/members/xieyuting/HBA/differential analysis of brain regions_20230111")
info<-read.csv("FIleInfo_HBA_20220816_withoutFALSE_and_pool.csv",header = T,sep = ",",stringsAsFactors = F)
load(file = "matrix_G.RData")

info$Lobe_BBA <- paste0(info$Lobe, sep = "_", info$BBA_Anatomical.and.modified.Cyto.architectonic.descriptions)
# load(file = "matrix_W.RData")
datExpr0 <- dfG
datExpr0$Lobe_BBA <- info$Lobe_BBA[match(rownames(datExpr0), info$sample)]
datExpr1 <- datExpr0[, -(ncol(datExpr0)-1)]

region <- unique(datExpr1$Lobe_BBA)

datExpr2<-reshape2::melt(datExpr1,id="Lobe_BBA")
datExpr3<-reshape2::dcast(datExpr2, Lobe_BBA~variable, mean, na.rm=T)
datExpr3 <- data.frame(datExpr3, row.names = "Lobe_BBA")

library(foreach)
library(doParallel)
cl <- makeCluster(40)
registerDoParallel(cl)

FC<-foreach(i=1:length(region),.combine=rbind) %:%
    foreach(j=1:length(region),.combine=rbind) %dopar% {
    
    data1<-datExpr3[which(rownames(datExpr3)==region[i]),]
    data2<-datExpr3[which(rownames(datExpr3)==region[j]),]
    
    # data_mean1 <- apply(data1,2,function(x){mean(as.numeric(x))})
    # data_mean2 <- apply(data2,2,function(x){mean(as.numeric(x))})
    
    data_FC <- as.data.frame(data1-data2)
    
    rownames(data_FC)<-paste(region[i],region[j],sep="_X_")
    
    return(data_FC)
  }
#40; 20:53~21:58
abs_log2_FC<-abs(FC) %>% as.data.frame()
# log2_FC<-FC %>% as.data.frame()

stopImplicitCluster()
stopCluster(cl)

setwd("W:/members/xieyuting/HBA/differential analysis of brain regions_20230111")
save(abs_log2_FC,file="G_matter_abs_log2_FC.RData")

# Na.ratio<-foreach(i=1:length(region),.combine=c) %dopar% {
#   
#   D1<-df9[which(df9$region==region[i]),-ncol(df9)]
#   Na<-round(sum(is.na(D1))/(ncol(D1)*nrow(D1)),digits = 2)
#   
#   
#   return(Na)
#   
# }
# Na.ratio<-data.frame(ratio_Na=Na.ratio)
# rownames(Na.ratio)<-region

# df9[is.na(df9)]<-min(df9[,-ncol(df9)],na.rm=T)/2
# df9<-df9[,-ncol(df9)] %>% log2(.) #几何平均值
# df9$region<-df8$region
```

#t-test
```{r}
library(foreach)
library(doParallel)
cl <- makeCluster(40)
registerDoParallel(cl)

# log2_df9<-log(df9[,-ncol(df9)],base=2)
# log2_df9$region<-df9$region
# log2_df9<-df9

t<-foreach(i=1:length(region),.combine=rbind) %:%
        foreach(j=1:length(region),.combine=rbind) %dopar% {
    
        t1<-datExpr1[c(which(datExpr1$Lobe_BBA==region[i]),which(datExpr1$Lobe_BBA==region[j])),]
        
        pval<-sapply(c(1:(ncol(t1)-1)),function(x){
              p<-try(t.test(t1[,x]~t1$Lobe_BBA),silent=TRUE)
              if(is(p,"try-error")) {NA} else {p$p.value}
              })
        
        t2 <- t(as.data.frame(pval))
        rownames(t2)<-paste(region[i],region[j],sep="_X_")
        
        return(t2)
  }
#20:58~21:50
t<-as.data.frame(t)
colnames(t)<-colnames(datExpr1)[-ncol(datExpr1)]

adj_p<-as.data.frame(t(apply(t,1,function(x){p.adjust(x,method = "BH")})))

stopImplicitCluster()
stopCluster(cl)

setwd("W:/members/xieyuting/HBA/differential analysis of brain regions_20230111")
save(adj_p,file="W_matter_T_test_adj_P.RData")
# setwd("Z:/project/HBA_xiaoqi/DataAnalysis20220712/differential analysis of brain regions based on dysregulated proteins")
# write.csv(adj_p,"adjust_pvalue_NO_NA.csv")
```

#Selectproteins:  BH-corrected-P < 0.05
```{r}
setwd("Z:/members/xieyuting/HBA/differential analysis of brain regions_20230111")
load(file="Protein_G_matter_abs_log2_FC.RData")
load(file="Protein_G_matter_T_test_adj_P.RData")

library(foreach)
library(doParallel)
cl <- makeCluster(3)
registerDoParallel(cl)

#log2(FC) > 1.5 | log2(FC) < -1.5
adj_p1 <- adj_p[rownames(abs_log2_FC),colnames(abs_log2_FC)]
adj_p1[is.na(adj_p1)]<-2

DP<-foreach(i=1:nrow(adj_p1),.combine=rbind,.inorder=TRUE) %dopar% {
  a<-adj_p1[i,]
  b<-abs_log2_FC[i,]
  data<-c()
  data[which((a < 0.05) & (b > 1.5))]<- 1
  data[which(a >= 0.05 | b <= 1.5)]<- 0   #log2(FC) < -1.5
  return(data)
}

DP<-as.data.frame(DP)
rownames(DP)<-rownames(adj_p1)
colnames(DP)<-colnames(adj_p1)


stopImplicitCluster()
stopCluster(cl)

# setwd("Z:/project/HBA_xiaoqi/DataAnalysis20220712/differential analysis of brain regions based on dysregulated proteins")
# write.csv(DP,"differential_protiens_abs_log2_FC_1.5_adj_P_0.05_no_NA.csv")
```

#heatmap
```{r}
HP<-as.data.frame(apply(DP,1,function(x){sum(x==1)}))
colnames(HP)<-"protein_number"
HP$V1<-rownames(HP)
HP$V2<-rownames(HP)
HP$V1<- sapply(HP$V1,function(x){str_split(x,"_X_")[[1]][1]})
HP$V2<- sapply(HP$V2,function(x){str_split(x,"_X_")[[1]][2]})
HP$value<-HP$protein_number
HP<-HP[,-1]
HP_data<-acast(HP,V1~V2)

# Na.ratio$a<-rep(1,98)
# Na.ratio1<-Na.ratio[rownames(HP_data),]
# rownames(HP_data)<-paste(colnames(HP_data),Na.ratio1$ratio_Na,sep="--NA_ratio-")

setwd("D:/1_guomics_xyt/1Project/hba_research_workflow/9_DataAnalysis_20220901/differential analysis of brain regions_20230111")
# save(DP, HP, HP_data, file="G_matter_differentially expressed proteins data_DP_HP_HP_data.RData")
load(file="G_matter_differentially expressed proteins data_DP_HP_HP_data.RData")
# setwd("W:/members/xieyuting/HBA/differential analysis of brain regions_20230111")
# info<-read.csv("FIleInfo_HBA_20220816_withoutFALSE_and_pool.csv",header = T,sep = ",",stringsAsFactors = F)
# info$Lobe_BBA <- paste0(info$BBA_Anatomical.and.modified.Cyto.architectonic.descriptions, seq="--", info$Lobe)
# info_order <- data.frame(BBA=info$BBA_Anatomical.and.modified.Cyto.architectonic.descriptions, Lobe_BBA=info$Lobe_BBA, order=info$Order, gyrus=info$Gyrus)
# info_order <- unique(info_order)
# info_order <- info_order[-which(info_order$BBA=="#N/A"), ]
# info_order <- info_order[order(info_order$order), ]
# info_order$Lobe <-  sapply(info_order$Lobe_BBA,function(x){str_split(x,"--")[[1]][2]})
# info_order$ABB <- sapply(info_order$BBA,function(x){str_split(x,",")[[1]][1]})
# rownames(info_order) <- info_order$BBA
# setwd("D:/1_guomics_xyt/1Project/hba_research_workflow/9_DataAnalysis_20220901/differential analysis of brain regions_20230111")
# # save(DP, HP, HP_data, file="G_matter_differentially expressed proteins data_DP_HP_HP_data.RData")
# write.csv(info_order, "info_order_of_BBA.csv")
info_order <- read.csv("info_order_of_BBA.csv")

names <- data.frame(name=rownames(HP_data))
names$v1 <- sapply(names$name,function(x){str_split(x,"_")[[1]][1]})
names$v2 <- sapply(names$name,function(x){str_split(x,"_")[[1]][2]})
names$v3 <- sapply(names$name,function(x){str_split(x,"_")[[1]][3]})
names$v4 <- sapply(names$name,function(x){str_split(x,"_")[[1]][4]})
# names2 <- names[!is.na(names$v3), ]
names$v5 <- paste0(names$v2, seq="_", names$v3, seq="_", names$v4)
names$v6 <- sapply(names$v5,function(x){gsub("_NA","", x)})
names$BBA_Lobe <- paste0(names$v6, seq="--", names$v1)

rownames(HP_data) <- names$BBA_Lobe[match(rownames(HP_data), names$name)]
colnames(HP_data) <- rownames(HP_data)
HP_data <- HP_data[info_order$Lobe_BBA, info_order$Lobe_BBA]
HP_data[HP_data==0]<-NA
# HP_data<-HP_data+1
log_HPdata<-log2(HP_data)
log_HPdata[is.na(log_HPdata)]<-0
rownames(log_HPdata) <- info_order$ABB[match(rownames(log_HPdata), info_order$Lobe_BBA)]
colnames(log_HPdata) <- rownames(log_HPdata)

anno <- data.frame(Lobe=info_order$Lobe, gyrus=info_order$gyrus)
rownames(anno) <- info_order$ABB

# colors <- data.frame(group=c(unique(info_order$Lobe), unique(info_order$gyrus)))
# setwd("D:/1_guomics_xyt/1Project/hba_research_workflow/9_DataAnalysis_20220901/differential analysis of brain regions_20230111")
# write.csv(colors, "colors.csv")
setwd("D:/1_guomics_xyt/1Project/hba_research_workflow/9_DataAnalysis_20220901/differential analysis of brain regions_20230111")
colors <- read.csv("colors.csv")
colors <- colors[, -1]

lobe_col <- colors[match(unique(anno$Lobe), colors$group), ]
lobe_col1 <- lobe_col$colors
names(lobe_col1) <- lobe_col$group

gyrus_col <- colors[match(unique(anno$gyrus), colors$group), ]
gyrus_col1 <- gyrus_col$colors
names(gyrus_col1) <- gyrus_col$group

ann_colors = list( Lobe=lobe_col1, gyrus=gyrus_col1)
# ann_colors <- list(Lobe=c(FrontalLobe= "#9a8fa2", TemporalLobe= "#98a8be"),
#                    gyrus=c(SuperiorFrontalGyrus_SFG="#a2d2af", MiddleFrontalGyrus_MFG="#c5dd9d"))

# brewer.pal(n = 8, name = "Spectral")

setwd("D:/1_guomics_xyt/1Project/hba_research_workflow/9_DataAnalysis_20220901/differential analysis of brain regions_20230111")
library(pheatmap)
library(RColorBrewer)
pheatmap(log_HPdata,
         fontsize=15,
         color = colorRampPalette(c("white","#6A7F67","firebrick4"))(500), 
         annotation_col = anno,
         # annotation_col = gyrus,
         # annotation_row = anno,
         # annotation_row = gyrus,
         annotation_colors = ann_colors,
         #na_color = "gray",
         trace = "none", 
         scale = "none", 
         cluster_rows = F,
         cluster_cols = F,
         show_rownames = T, 
         show_colnames = T, 
         filename = "G_matter_differential_protein_numbers_AbsLog2FC_1.5_121_Regions_log2_number_annotation.pdf",       
         main = "Number of differentially expressed proteins between pairs of brain regions in grey matter",
         height = 40, width = 44)


```



```{r}
stopImplicitCluster()
stopCluster(cl)
```
